moved code to individual source files for (hopefully) easier modification.
