+++ data/states/keno/main.py
@@ -1,290 +1,21 @@
 import os
-import random
 import pygame as pg
 from ...components.loggable import getLogger
 from ...components.warning_window import NoticeWindow
-from ...components.labels import Label, MultiLineLabel, NeonButton, ButtonGroup
+from ...components.labels import Label, ButtonGroup, NeonButton
 from ... import tools, prepare
+from .keno_card import KenoCard
+from .pay_table import PayTable
+from .round_history import RoundHistory
+from .action import Action
 from .model import Wallet, Pot, InsufficientFundsException
+from .helpers import pick_numbers, PAYTABLE
 
 # Utilize the logger along with the following functions to print to console instead of prints.
 log = getLogger("KENO")
 #log.debug("testing debug log")
 #log.info("testing info log")
 #log.error("testing error log")
-
-#http://casinogamblingtips.info/tag/pay-table
-PAYTABLE = [
-    [(0, 0)], #0
-    [(1, 3)], #1
-    [(2, 12)], #2
-    [(2, 1), (3, 40)], #3
-    [(2, 1), (3, 2), (4, 120)], #4
-    [(3, 1), (4, 18), (5, 800)], #5
-    [(3, 1), (4, 3), (5, 80), (6, 1500)], #6
-    [(4, 1), (5, 18), (6, 360), (7, 5000)], #7
-    [(5, 10), (6, 75), (7, 1000), (8, 15000)], #8
-    [(5, 4), (6, 35), (7, 250), (8, 3000), (9, 20000)], #9
-    [(5, 2), (6, 15), (7, 100), (8, 1500), (9, 8000), (10, 25000)], #10
-]
-
-def pick_numbers(spot):
-    numbers = []
-    while len(numbers) < spot:
-        number = random.randint(0, 79)
-        if number not in numbers:
-            numbers.append(number)
-    return numbers
-
-def is_winner(spot, hit):
-    paytable = PAYTABLE[spot]
-    for entry in paytable:
-        if entry[0] == hit:
-            return True
-    return False
-
-class RoundHistory(object):
-    '''Round history showing hits per round.'''
-    def __init__(self, card):
-        self.rect = pg.Rect(24, 200, 304, 554)
-        self.font = prepare.FONTS["Saniretro"]
-        self.color = '#181818'
-        self.card = card
-
-        self.header_labels = []
-        self.header_labels.extend([Label(self.font, 32, 'ROUND', 'white', {'center':(100,224)})])
-        self.header_labels.extend([Label(self.font, 32, 'HITS', 'white', {'center':(280,224)})])
-
-        self.result_labels = []
-
-        self.round_x = 100
-        self.hit_x   = 280
-        self.row_y   = 224+32
-
-        self.rounds  = 1
-
-    def clear(self):
-        self.rounds = 1
-        self.result_labels = []
-        self.row_y = 224+32
-
-    def update(self, spot, hits):
-        if self.rounds % 17 == 0:
-            self.clear()
-
-        color = "white"
-
-        if is_winner(spot, hits):
-            color = "gold3"
-
-        self.result_labels.extend([Label(self.font, 32, str(self.rounds), color, {'center':(self.round_x, self.row_y)})])
-        self.result_labels.extend([Label(self.font, 32, str(hits), color, {'center':(self.hit_x, self.row_y)})])
-        self.row_y+=32
-        self.rounds+=1
-
-    def draw(self, surface):
-        pg.draw.rect(surface, pg.Color(self.color), self.rect, 0)
-
-        for label in self.header_labels:
-            label.draw(surface)
-
-        for label in self.result_labels:
-            label.draw(surface)
-
-class PayTable(object):
-    '''Paytable readout for desired spot count'''
-    def __init__(self, card):
-        self.rect = pg.Rect(1036, 200, 340, 554)
-        self.font = prepare.FONTS["Saniretro"]
-        self.color = '#181818'
-        self.card = card
-
-        self.header_labels = []
-        self.header_labels.extend([Label(self.font, 32, 'HIT', 'white', {'center':(1080,224)})])
-        self.header_labels.extend([Label(self.font, 32, 'WIN', 'white', {'center':(1280,224)})])
-
-        self.pay_labels = []
-
-    def update(self, spot, bet=1):
-        self.pay_labels = []
-        row = PAYTABLE[spot]
-        hit_x = 1080
-        win_x = 1280
-        row_y = 224+32
-        for entry in row:
-            hit, win = entry
-            win *= bet
-            self.pay_labels.extend([Label(self.font, 32, str(hit), 'white', {'center':(hit_x, row_y)})])
-            self.pay_labels.extend([Label(self.font, 32, str(win), 'white', {'center':(win_x, row_y)})])
-            row_y+=32
-
-
-    def draw(self, surface):
-        pg.draw.rect(surface, pg.Color(self.color), self.rect, 0)
-
-        for label in self.header_labels:
-            label.draw(surface)
-
-        for label in self.pay_labels:
-            label.draw(surface)
-
-class KenoSpot(object):
-    COLORS = {
-        'open': '#181818',
-        'owned': '#9c8013',
-        'hit': '#eedb1e',
-        'miss': '#690808',
-    }
-
-    """A spot on a Keno card."""
-    def __init__(self, left, top, width, height, label, img=None):
-        self.rect = pg.Rect(left, top, width, height)
-        label.rect.center = self.rect.center
-        self.label = label
-        self.color = self.COLORS['open']
-        self.img = img
-
-        self.owned = False
-        self.hit   = False
-
-    def reset(self):
-        self.owned = False
-        self.hit   = False
-        self.update_color()
-
-    def toggle_owned(self):
-        self.owned = not self.owned
-        self.update_color()
-
-    def toggle_hit(self):
-        self.hit = not self.hit
-        self.update_color()
-
-    def update_color(self):
-        if self.owned:
-            self.color = self.COLORS['owned']
-        else:
-            self.color = self.COLORS['open']
-
-        if self.hit and self.owned:
-            self.color = self.COLORS['hit']
-        elif self.hit:
-            self.color = self.COLORS['miss']
-
-    def draw(self, surface):
-        pg.draw.rect(surface, pg.Color(self.color), self.rect, 0)
-
-        if self.img:
-            if self.hit:
-                surface.blit(self.img, self.rect)
-            else:
-                self.label.draw(surface)
-        else:
-            self.label.draw(surface)
-
-class KenoCard(object):
-    def __init__(self, sprite_sheet=None):
-        self.font = prepare.FONTS["Saniretro"]
-        self.sheet = sprite_sheet
-        self.spots = []
-        self.current_pick = []
-        self.build()
-
-    @property
-    def spot_count(self):
-        count = 0
-        for spot in self.spots:
-            if spot.owned:
-                count+=1
-        return count
-
-    @property
-    def hit_count(self):
-        count = 0
-        for spot in self.spots:
-            if spot.hit and spot.owned:
-                count+=1
-        return count
-
-    def build(self):
-        font_size = 48
-        text = "0"
-        text_color = "white"
-        rect_attrib = {'center':(0,0)}
-
-        x_origin = 336
-        x = x_origin
-        y = 200
-        for row in range(0,8):
-            for col in range(1,11):
-                text = str(col+(10*row))
-                label = Label(self.font, font_size, text, text_color, rect_attrib)
-
-                if self.sheet:
-                    spot = KenoSpot(x, y, 64, 64, label, self.sheet[int(text)-1])
-                else:
-                    spot = KenoSpot(x, y, 64, 64, label)
-
-                self.spots.extend([spot])
-                x += 70
-            y += 70
-            x = x_origin
-
-    def toggle_owned(self, number):
-        self.spots[number].toggle_owned()
-
-    def toggle_hit(self, number):
-        self.spots[number].toggle_hit()
-
-    def ready_play(self, clear_all=False):
-        for spot in self.spots:
-            spot.hit = False
-            spot.update_color()
-
-        if clear_all:
-            for spot in self.spots:
-                spot.owned = False
-                spot.update_color()
-
-    def reset(self):
-        for spot in self.spots:
-            spot.reset()
-
-    def update(self, mouse_pos):
-        for spot in self.spots:
-            if spot.rect.collidepoint(mouse_pos):
-                if (self.spot_count < 10 and not spot.owned) or spot.owned:
-                    spot.toggle_owned()
-
-    def draw(self, surface):
-        x_pos = 64
-        y_pos = 136
-        size  = 64
-        for ball in self.current_pick:
-            img = self.spots[ball].img
-            surface.blit(img, pg.Rect(x_pos, y_pos, size, size))
-            x_pos += size
-
-
-        for spot in self.spots:
-            spot.draw(surface)
-
-class Action(object):
-    def __init__(self, rect, label, callback):
-        self.rect = rect
-        self.label = label
-        self.label.rect.center = self.rect.center
-        self.color = '#181818'
-        self.callback = callback
-        
-    def execute(self, mouse_pos):
-        if self.callback:
-            if self.rect.collidepoint(mouse_pos):
-                self.callback()
-            
-    def draw(self, surface):
-        pg.draw.rect(surface, pg.Color(self.color), self.rect, 0)
-        self.label.draw(surface)
 
 class Keno(tools._State):
     """Class to represent a casino game."""
